size to be int64 only - shape dtype

+++ onnx/backend/test/case/node/affinegrid.py
@@ -71,7 +71,7 @@ class AffineGrid(Base):
             test_name = "test_affine_grid_2d"
             if align_corners == 1:
                 test_name += "_align_corners"
-            expect(node, inputs=[theta_2d, np.array([N, C, W, H])], outputs=[grid], name=test_name)
+            expect(node, inputs=[theta_2d, np.array([N, C, W, H], dtype=np.int64)], outputs=[grid], name=test_name)
 
     @staticmethod
     def export_3d() -> None:
@@ -104,4 +104,4 @@ class AffineGrid(Base):
             test_name = "test_affine_grid_3d"
             if align_corners == 1:
                 test_name += "_align_corners"
-            expect(node, inputs=[theta_3d, np.array([N, C, D, W, H])], outputs=[grid], name=test_name)
+            expect(node, inputs=[theta_3d, np.array([N, C, D, W, H], dtype=np.int64)], outputs=[grid], name=test_name)
